Split `tests/client/rpcSessions`
